ECOM-15903: Change order item id type to Long in OrderForCalculate and CalculateOrderDetailsResult classes.

+++ src/test/kotlin/com/ecwid/apiclient/v3/entity/CartsTest.kt
@@ -477,7 +477,7 @@ class CartsTest : BaseEntityTest() {
 
 	private fun generateTestItem(): OrderForCalculate.OrderItem {
 		return OrderForCalculate.OrderItem(
-			id = randomId(),
+			id = randomId().toLong(),
 			productId = randomId(),
 			categoryId = randomId(),
 			price = 22.2,
